Replace mocks/stubs in tests with LSP
Previously, the tests were tied to the functions that our Node language
server exposes.
This changes the tests to operate as an LSP Client.
While clients are a bit low level, it means that the tests are no longer
coupled to the values that `vscode-languageserver/node` expects or mocking
and stubbing.

This also also fix coverage, which wasn’t reported due to typo where
`c8 ignore end` was used, which doesn’t do anything, instead of
`c8 ignore stop`.

It also adds stack traces when plugins crash, without them it’s hard to
figure out what happened where and why.