Add support for workspaces
This adds support for workspaces, which in most cases will be just a simple
single workspace (sort of like a working directory on the terminal), but in
more complex cases can be multiple workspaces and which change over time.

The closest open workspace is chosen when processing a file, in batches of
files that belong together in a workspace.

For backwards compatibility with LSP 3.6.0, `event.rootUri` is also supported.

Closes GH-20.